Correct test inference API test
